requirements.txt
django
djangorestframework
django-cors-headers
django-debug-toolbar
django-extensions
python-dotenv
drf-spectacular
django-unfold
django-import-export
psycopg